requirements.txt
# Flask - Web Framework
Flask==0.11.1

# pymongo - MongoDB Driver for python
pymongo==3.2.2

# jinja2 - Template tools
Jinja2==2.8

# inflect - Generate plurals
inflect==0.2.5

# pycrypto - Encrypt and decrypt
PyCrypto==2.6.1
